move getPrices to a utils file
